cli: Improve import error output (#46269)
To be more user-friendly, we catch some errors that may indicate "you
have to run `pnpm install`" and display helpful information to the user.
But we're still dumping the whole error object, which is generally not
terribly useful because all the stack trace refers to "files" like
`node:internal/modules/esm/loader`.

Instead, when it's one of the errors we recognize, print only the
message without the trace, and do that in a dimmer color if possible.

And also fix the bolding of our message, which was probably broken
by #40786.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/20142874929

Upstream-Ref: Automattic/jetpack@041c012de2c04e7ebe65fc3320edcb8118f793f8